requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-crontab==0.7.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
